Navbar changes; remove some routes
