+++ controllers/diners.controller.js
@@ -4,54 +4,6 @@ var bcrypt = require('bcryptjs');
 var { check, validationResult } = require('express-validator');
 var passport = require('passport');
 
-exports.showAllDiners = async (req, res, next) => {
-  try {
-    const diners = await db.any('SELECT * FROM Diners');
-    res.render('diners', {
-      title: 'Diners',
-      diners: diners
-    });
-  } catch (e) {
-    next(e);
-  }
-};
-
-exports.showDinerProfile = async (req, res, next) => {
-  try {
-    const diner = db.one('SELECT * FROM Diners NATURAL JOIN Users WHERE uname=$1', [
-      req.params.uname
-    ]);
-    const points = db.one(
-      'SELECT COUNT(*) FROM ReserveTimeslots WHERE duname=$1',
-      [req.params.uname]
-    );
-    const mostVisited = db.any(
-      'SELECT rname, raddress FROM ReserveTimeslots WHERE duname=$1 GROUP BY rname, raddress ORDER BY count(*) DESC LIMIT 3',
-        [req.params.uname]
-    );
-    const reviews = db.any(
-      'SELECT rname, rating, review FROM ReserveTimeslots WHERE duname=$1 ORDER BY r_date DESC, r_time DESC LIMIT 3',
-        [req.params.uname]
-    );
-    const history = db.any(
-      "SELECT r_date, to_char(r_date, 'DD MON YYYY') AS date, r_time, to_char(r_time, 'HH12.MIPM') AS time, rname, raddress FROM ReserveTimeslots WHERE duname=$1 ORDER BY r_date DESC, r_time DESC LIMIT 3",
-        [req.params.uname]
-    );
-	Promise.all([diner, points, mostVisited, reviews, history]).then(values => {
-    res.render('diner', {
-      title: values[0].uname,
-      diner: values[0],
-      points: values[1],
-      visited: values[2],
-      reviews: values[3],
-      history: values[4]
-    });
-	})
-  } catch (e) {
-    next(e);
-  }
-};
-
 exports.showReservations = async (req, res, next) => {
   try {
     const reservations = db.any(
