+++ views/components/navbar.pug
@@ -1,5 +1,5 @@
 mixin navbar
-  nav.navbar.navbar-expand-xl.navbar-light.fixed-top.bg-transparent
+  nav.navbar.navbar-expand-xl.navbar-light.position-absolute.bg-transparent#navbar
     a.navbar-brand(href='/home') 
       img.d-inline-block.align-right(src='/images/logo.png' width='60' height='40')
     //Ketchup & Chillᴉ
