+++ views/message.pug
@@ -1,4 +1,4 @@
 .messages
   each type in Object.keys(messages)
     each message in messages[type]
-      div(class='alert alert-' + type) #{message}
+      div(id='alert' class='alert alert-' + type) #{message}
